feat(bgp): negotiate capabilities and carry MP-BGP EVPN over the session
The speaker could only ever describe IPv4 unicast: OPEN carried no optional
parameters, the AS_PATH held 16-bit ASNs, and the UPDATE codec had no way to
express a family whose NLRI is not a list of IPv4 prefixes.

Three pieces, and they only mean something together:

  * bgp_caps  - RFC 5492 capability framework. Both OPENs are intersected into
    a negotiated AFI/SAFI set, so whether a session may carry EVPN is a single
    lookup rather than an assumption. A speaker offering no Multiprotocol
    capability is a legacy one and negotiates IPv4 Unicast alone, which is what
    keeps a plain RFC 4271 session working unchanged.

  * 4-octet ASNs - AsPath now holds u32, and the encoding width comes from the
    negotiation rather than being fixed. Where a 2-octet field cannot hold the
    value it gets AS_TRANS and the truth travels in the Four-Octet AS
    capability or AS4_PATH (RFC 6793), never a truncation to a different real
    AS. The width is passed into the parser instead of guessed, because the
    bytes genuinely do not say: a two-ASN wide segment and a four-ASN narrow
    one are the same length and both decode.

  * bgp_mp / bgp_evpn - MP_REACH_NLRI and MP_UNREACH_NLRI, and the EVPN RIBs
    behind them. Which path attributes are mandatory now depends on what the
    UPDATE announces: IPv4 NLRI still needs NEXT_HOP, an MP_REACH does not
    because the family carries its own, and a pure withdrawal needs nothing.

The EVPN Loc-RIB is recomputed from the Adj-RIB-In tables rather than patched,
so a withdrawn route, a lost session and a moved host are the same operation -
the input set changed - and none of them can leave state behind. Route Target
import filters on the way in, so a route no local instance asked for never
reaches the Loc-RIB at all. MAC Mobility is compared ahead of every ordinary
attribute, or a lower peer address could pin traffic to the location a host
has left.

Two defects fixed along the way, both reachable from the wire:

  * EvpnNlri::parse indexed past the end of a MAC/IP route that claimed a
    32-bit host IP in a body too short to hold one. The length check ran before
    the variable-length field it was supposed to bound.
  * The MAC and IP length fields were read and then ignored, so a route with a
    non-48-bit MAC or a 128-bit IP shifted every field after it and was decoded
    as if it had not.

Attribute encoding also picks the extended-length form past 255 bytes; an EVPN
UPDATE carrying many routes would otherwise have written its length as a u8.

+++ src/bgp_rib.rs
@@ -45,7 +45,7 @@ pub struct BgpPath {
     pub source: PathSource,
     /// Address of the peer that advertised it; unspecified for locally originated paths.
     pub peer_addr: Ipv4Address,
-    pub peer_as: u16,
+    pub peer_as: u32,
     pub peer_router_id: Ipv4Address,
     pub origin: BgpOrigin,
     pub as_path: AsPath,
@@ -522,12 +522,12 @@ impl RoutePolicy {
 mod tests {
     use super::*;
 
-    fn path(peer: u8, local_pref: u32, as_path: Vec<u16>, origin: BgpOrigin) -> BgpPath {
+    fn path(peer: u8, local_pref: u32, as_path: Vec<u32>, origin: BgpOrigin) -> BgpPath {
         BgpPath {
             prefix: Ipv4Prefix::new(Ipv4Address::new(10, 30, 0, 0), 24),
             source: PathSource::Ebgp,
             peer_addr: Ipv4Address::new(10, 0, 0, peer),
-            peer_as: 65000 + peer as u16,
+            peer_as: 65000 + peer as u32,
             peer_router_id: Ipv4Address::new(1, 1, 1, peer),
             origin,
             as_path: AsPath::sequence(as_path),
